feat(Gravity System): Implement the collisions when applying gravity
Since the gravity is the only movement system that make player collide the ground, I implemented the collision here. With more time, I would have implemented a checkCollisionHandler and a MoveEntity3D handler. Then, if the gravity sends an event to the collider handler, if it collides, doesn't apply the movement, if not, call the moveEntity3D handler